+++ controllers/User.ts
@@ -1,10 +1,13 @@
 const { bcryptjs, generateToken } = require("../services/Auth");
-const Users = require("../models/UserModel");
+import Users from "../models/UserModel";
 const Token = require("../models/TokenModel");
 const Product = require("../models/ProductModel");
 const Orders = require("../models/OrderModel");
+import { Request, Response } from "express";
+import { AuthenticatedRequest } from "../interfaces/Users";
+import { LoginBody } from "../interfaces/Users";
 
-async function signUpForm(req, res) {
+async function signUpForm(req: Request, res: Response) {
   const hashPassword = await bcryptjs.hash(req.body.password, 10);
   console.log(hashPassword);
   const newUser = {
@@ -20,7 +23,7 @@ async function signUpForm(req, res) {
   res.status(201).json({ message: "User created...!" });
 }
 
-async function loginForm(req, res) {
+async function loginForm(req: Request<{}, {}, LoginBody>, res: Response) {
   const user = await Users.findOne({ userEmail: req.body.email });
   if (!user) {
     console.log("User not found");
@@ -32,7 +35,7 @@ async function loginForm(req, res) {
   const isMatch = await bcryptjs.compare(req.body.password, user.userPassword);
   if (!isMatch) {
     console.log("Invalid credentials");
-    return res.status(400).send("Invalid credentials");
+    return res.status(400).json({ message: "Invalid credentials" });
   }
 
   console.log(user.userPassword);
@@ -40,13 +43,17 @@ async function loginForm(req, res) {
   res.cookie("refreshToken", refreshToken, {
     httpOnly: true,
     secure: true,
-    sameSite: "Strict",
+    sameSite: "strict",
     path: "/refreshToken",
   });
   res.json({ token: accessToken });
 }
 
-async function userProfile(req, res) {
+async function userProfile(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
+
   if (req.user.role === "user") {
     if (req.user.status === "blocked") {
       res.status(200).json({
@@ -63,7 +70,11 @@ async function userProfile(req, res) {
   }
 }
 
-async function updateStatus(req, res) {
+async function updateStatus(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
+
   try {
     if (req.user.role === "admin") {
       const adminId = await Users.findOne({ userRole: "admin" }, { _id: 1 });
@@ -96,7 +107,10 @@ async function updateStatus(req, res) {
   }
 }
 
-async function deleteUser(req, res) {
+async function deleteUser(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
   try {
     if (req.user.role === "admin") {
       await Users.findByIdAndDelete(req.body.id);
@@ -109,7 +123,10 @@ async function deleteUser(req, res) {
   }
 }
 
-async function logout(req, res) {
+async function logout(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
   try {
     const user = await Users.findOne({ _id: req.body.id });
     const token = req.token;
@@ -134,7 +151,10 @@ async function logout(req, res) {
   }
 }
 
-async function dashboard(req, res) {
+async function dashboard(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
   try {
     if (req.user.role === "admin") {
       const allProducts = await Product.find();
@@ -163,7 +183,10 @@ async function dashboard(req, res) {
   }
 }
 
-async function adminProfile(req, res) {
+async function adminProfile(req: AuthenticatedRequest, res: Response) {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized. No user found." });
+  }
   if (req.user.role === "admin") {
     res
       .status(200)
